Renamed GRexPersister in :grex-android module to GRexAndroidPersister. Renamed GRexCorePersister to GRexPersister and moved out of core package to grex package.
